Refactored follower and followee count to be in the follower and followee services, respectively
