+++ app/src/main/java/edu/byu/cs/tweeter/Client/model/services/LoginServiceProxy.java
@@ -22,8 +22,7 @@ import byu.edu.cs.tweeter.shared.response.StartUpResponse;
  */
 public class LoginServiceProxy extends Service implements LoginService {
     private static final String LOGIN_URL_PATH = "/login";
-    private static final String FOLLOWER_COUNT_URL_PATH = "/followercount";
-    private static final String FOLLOWEE_COUNT_URL_PATH = "/followeecount";
+
     public LoginResponse checkLogin(LoginRequest request) throws IOException, TweeterRemoteException {
         return getServerFacade().checkLogin(request, LOGIN_URL_PATH);
     }
@@ -32,14 +31,5 @@ public class LoginServiceProxy extends Service implements LoginService {
 //    public StartUpResponse startUp(StartUpRequest request) {
 //        return getServerFacade().startUp(request);
 //    }
-    public FollowerCountResponse getFollowerCount(FollowerCountRequest request) throws IOException, TweeterRemoteException {
-        String param_alias = "/" + request.getUser().getAlias();
-        String newUrlPath = FOLLOWER_COUNT_URL_PATH + param_alias;
-        return getServerFacade().getFollowerCount(request, newUrlPath);
-    }
-    public FolloweeCountResponse getFolloweeCount(FolloweeCountRequest request) throws IOException, TweeterRemoteException {
-        String param_alias = "/" + request.getUser().getAlias();
-        String newUrlPath = FOLLOWEE_COUNT_URL_PATH + param_alias;
-        return getServerFacade().getFolloweeCount(request, newUrlPath);
-    }
+
 }
