+++ shared/src/main/java/byu/edu/cs/tweeter/shared/model/domain/service/FollowerService.java
@@ -3,7 +3,9 @@ package byu.edu.cs.tweeter.shared.model.domain.service;
 import java.io.IOException;
 
 import byu.edu.cs.tweeter.shared.net.TweeterRemoteException;
+import byu.edu.cs.tweeter.shared.request.FollowerCountRequest;
 import byu.edu.cs.tweeter.shared.request.FollowersRequest;
+import byu.edu.cs.tweeter.shared.response.FollowerCountResponse;
 import byu.edu.cs.tweeter.shared.response.FollowersResponse;
 
 /**
@@ -11,4 +13,5 @@ import byu.edu.cs.tweeter.shared.response.FollowersResponse;
  */
 public interface FollowerService {
     public FollowersResponse getFollowers(FollowersRequest request) throws IOException, TweeterRemoteException;
+    public FollowerCountResponse getFollowerCount(FollowerCountRequest request) throws IOException, TweeterRemoteException;
 }
