+++ shared/src/main/java/byu/edu/cs/tweeter/shared/model/domain/service/FollowingService.java
@@ -4,7 +4,9 @@ package byu.edu.cs.tweeter.shared.model.domain.service;
 import java.io.IOException;
 
 import byu.edu.cs.tweeter.shared.net.TweeterRemoteException;
+import byu.edu.cs.tweeter.shared.request.FolloweeCountRequest;
 import byu.edu.cs.tweeter.shared.request.FollowingRequest;
+import byu.edu.cs.tweeter.shared.response.FolloweeCountResponse;
 import byu.edu.cs.tweeter.shared.response.FollowingResponse;
 
 /**
@@ -12,6 +14,5 @@ import byu.edu.cs.tweeter.shared.response.FollowingResponse;
  */
 public interface FollowingService {
     public FollowingResponse getFollowees(FollowingRequest request) throws AssertionError, IOException, TweeterRemoteException;
-
-
+    public FolloweeCountResponse getFolloweeCount(FolloweeCountRequest request) throws IOException, TweeterRemoteException;
 }
