+++ shared/src/main/java/byu/edu/cs/tweeter/shared/model/domain/service/LoginService.java
@@ -19,6 +19,5 @@ import byu.edu.cs.tweeter.shared.response.StartUpResponse;
 public interface LoginService {
     public LoginResponse checkLogin(LoginRequest request) throws IOException, TweeterRemoteException;
     //public StartUpResponse startUp(StartUpRequest request);
-    public FollowerCountResponse getFollowerCount(FollowerCountRequest request) throws IOException, TweeterRemoteException;
-    public FolloweeCountResponse getFolloweeCount(FolloweeCountRequest request) throws IOException, TweeterRemoteException;
+
 }
